+++ v2/Include/Controllers/Display/eosColorFrameBuffer_DMA2D.h
@@ -17,18 +17,18 @@ namespace eos {
 			using Dma2d = htl::DMA2D_1;
 
 			Color::Pixel *_buffer;
-			const int _framePitch;
+			const int16_t _framePitch;
 
-			inline Color::Pixel *getPixelPtr(int x, int y) const { return &_buffer[(y * _framePitch) + x]; }
+			inline Color::Pixel *getPixelPtr(int16_t x, int16_t y) const { return &_buffer[(y * _framePitch) + x]; }
 
 		protected:
-			void put(int x, int y, Color color) override;
-            void fill(int x, int y, int width, int height, Color color) override;
-            void copy(int x, int y, int width, int height, const Color *colors, int colorPitch) override;
-            void copy(int x, int y, int width, int height, const void *colors, ColorFormat colorFormat, int colorPitch) override;
+			void put(int16_t x, int16_t y, Color color) override;
+            void fill(int16_t x, int16_t y, int16_t width, int16_t height, Color color) override;
+            void copy(int16_t x, int16_t y, int16_t width, int16_t height, const Color *colors, int16_t colorPitch) override;
+            void copy(int16_t x, int16_t y, int16_t width, int16_t height, const void *colors, ColorFormat colorFormat, int16_t colorPitch) override;
 
 		public:
-			ColorFrameBuffer_DMA2D(int width, int height, int pitch, DisplayOrientation orientation, void *buffer);
+			ColorFrameBuffer_DMA2D(int16_t width, int16_t height, int16_t pitch, DisplayOrientation orientation, void *buffer);
             void *getBuffer() const override;
 	};
 }
